refactor: clear libs dir
