+++ public/partial-publicationCharts.html
@@ -19,7 +19,7 @@
 
     <div align="center">
       <fieldset style="width: 300px; border-radius: 10px">
-        <legend>Chart Type:</legend>
+        <legend>Data Range:</legend>
 
         <div>
           <input type="radio" id="yearly" name="type" value="Yearly" checked onclick="changeChart('yearly')" />
