+++ public/tools/dbplayground.html
@@ -8,12 +8,12 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.2/css/all.min.css" crossorigin="anonymous" />
   <link href="https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700|Roboto" rel="stylesheet" />
-  <script src="./../js/libs/codemirror.js"></script>
-  <link rel="stylesheet" href="./../js/libs/codemirror.css" />
-  <script src="./../js/libs/sql.js"></script>
-  <script src="./../js/libs/show-hint.js"></script>
-  <script src="./../js/libs/sql-hint.js"></script>
-  <link rel="stylesheet" href="./../js/libs/hints.css" />
+  <script src="./../js/libs/codemirror/codemirror.js"></script>
+  <link rel="stylesheet" href="./../js/libs/codemirror/codemirror.css" />
+  <script src="./../js/libs/sql/sql.js"></script>
+  <script src="./../js/libs/sql/show-hint.js"></script>
+  <script src="./../js/libs/sql/sql-hint.js"></script>
+  <link rel="stylesheet" href="./../js/libs/sql/hints.css" />
   <link rel="stylesheet" href="./../css/dbplayground.css" />
   <style>
     .anchor {
